Added Pantry functionality along with scanner.

+++ app/src/main/java/sheridancollege/prog39402/stayfresh/Peter/PantryFragment.kt
@@ -1,4 +1,4 @@
-package sheridancollege.prog39402.stayfresh
+package sheridancollege.prog39402.stayfresh.Peter
 
 import android.os.Bundle
 import android.util.Log
@@ -7,7 +7,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
+import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.firebase.auth.FirebaseAuth
 import sheridancollege.prog39402.stayfresh.databinding.FragmentPantryBinding
 
 class PantryFragment : Fragment() {
@@ -28,7 +30,17 @@ class PantryFragment : Fragment() {
 
         viewModel = ViewModelProvider(this).get(PantryViewModel::class.java)
 
-        val userId = "KR9TcjSaEybq0UigkFB4aMkJhz63"
+        // Get the current user's ID from FirebaseAuth
+        val userId = FirebaseAuth.getInstance().currentUser?.uid
+        if (userId == null) {
+            Log.e("PantryFragment", "No user is logged in.")
+            // Handle the case where there is no logged-in user
+            return binding.root
+        }
+
+        pantryAdapter = PantryAdapter { foodId ->
+            viewModel.deleteFood(userId, foodId)
+        }
 
         pantryAdapter = PantryAdapter { foodId ->
             viewModel.deleteFood(userId, foodId)
@@ -50,7 +62,21 @@ class PantryFragment : Fragment() {
             Log.d("PantryFragment", "Items Loaded: ${items.size}")
         }
 
+
+
         return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.addItemButton?.setOnClickListener {
+            val action = ContentFragmentDirections.actionContentFragmentToAddFragment()
+            findNavController().navigate(action)
+        }
+
+
+
     }
 
     override fun onDestroyView() {
